requirements.txt
flask
joblib
scikit-learn
pandas
requests
# requirements.txt
python-dotenv

